refactor: 댓글 개수 카운팅 조건에 대댓글도 추가

+++ Api/src/main/java/com/dingdong/api/idcard/dto/IdCardDetailsDto.java
@@ -33,10 +33,10 @@ public class IdCardDetailsDto {
     private CharacterType characterType;
 
     @Schema(description = "댓글 개수")
-    private Long commentCount;
+    private int commentCount;
 
     public static IdCardDetailsDto of(
-            IdCard idCard, List<KeywordDto> keywordDtos, Long commentCount) {
+            IdCard idCard, List<KeywordDto> keywordDtos, int commentCount) {
         return IdCardDetailsDto.builder()
                 .idCardId(idCard.getId())
                 .userId(idCard.getUserInfo().getUserId())
